min_unet_model/requirements.txt
numpy
pillow
tensorflow
matplotlib